Automate django testing on docker compose
Running `docker-compose up` will now run all django tests before starting the server.
Because django depends on the postgres database to be running when it runs tests, the CMD command in `Dockerfile.backend` which previousuly started the server has been moved to `docker-compose.yml`. The alternative was to allow django to run tests on a separate sqllite database in memory, but running tests on a (test) postgres database will help to catch subtle database platform issues.