Do async right this time
Turns out I was full of shit when I made the last commit. That code was not written with best practices.

This code actually allows throwing Axios' errors out to the components calling the user hook functions.

I've written the async functions a couple different ways to showcase the different styles.