Convert Portfolio display to Material UI
This was a big commit due to going down the rabbit hole and fixing things. Here are the changes:
1. Portfolio page now renders with Material UI components. There is support for a background image (a sample one is provided).
2. Cleaned portfolio types (previously in Types.tsx). Moved to `jinxui/types/PortfolioTypes.ts` and added the types to `jinxui`. Extended TSectionData to make TSection to avoid copy-paste. Changed (section) name field in TSection to be optional.
3. Made jinxui/header/HeaderBar 'sticky' so it doesn't block content beneath it.
4. Added lastName to UserContext, and made the user hook load this by default. Added a 'name' field to userData so you don't have to string concat everytime (see return of useUser).
5. Fixed some useUser functions (like getAccountDetails) to return data rather than an axios response object.

Please keep in mind that ideally no useUser functions should return anything HTTP/Axios response related.

The portfolio page renders sections with a generic Section component. This handles any sections that are a combination of a (section) name, text content, and image.
You're welcome to make other section components as you please. In order to add this to the display page, find Portfolio.tsx/SectionGrid and change the map function (which takes a list of section data, and converts them to the relevant section components (currently just the generic Section)).